requirements.txt
streamlit==1.38.0
pandas==2.2.2
python-pptx==0.6.23
numpy==1.26.4
Pillow==10.4.0
